Enable WooCommerce scripts if plugin is installed
